requirements.txt
arduino
pyserial
